requirements.txt
asgiref==3.7.2
Authlib==1.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
decorator==5.1.1
dj-database-url==2.1.0
Django==4.2.5
django-heroku==0.3.1
django-storages==1.13.1
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
Naked==0.1.32
packaging==23.1
Pillow==10.0.1
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycparser==2.21
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
self==2020.12.3
sqlparse==0.4.4
status==0.2.5
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
whitenoise==6.5.0
mysqlclient
itsdangerous
python-dotenv